addressing review and warnings

+++ interface/src/workload/GameWorkloadRenderer.cpp
@@ -9,6 +9,7 @@
 //
 #include "GameWorkloadRenderer.h"
 
+#include <cstring>
 #include <gpu/Context.h>
 
 
@@ -28,7 +29,7 @@ void GameSpaceToRender::run(const workload::WorkloadContextPointer& runContext, 
     if (!render::Item::isValidID(_spaceRenderItemID)) {
         _spaceRenderItemID = scene->allocateID();
         auto renderItem = std::make_shared<GameWorkloadRenderItem>();
-        renderItem->editBound().expandedContains(glm::vec3(0.0), 32000.0);
+        renderItem->editBound().expandedContains(glm::vec3(0.0f), 32000.0f);
         transaction.resetItem(_spaceRenderItemID, std::make_shared<GameWorkloadRenderItem::Payload>(std::make_shared<GameWorkloadRenderItem>()));
         scene->enqueueTransaction(transaction);
     }
@@ -41,7 +42,7 @@ void GameSpaceToRender::run(const workload::WorkloadContextPointer& runContext, 
 
     std::vector<workload::Space::Proxy> proxies(space->getNumAllocatedProxies());
 
-    space->copyProxyValues(proxies.data(), proxies.size());
+    space->copyProxyValues(proxies.data(), (uint32_t) proxies.size());
 
     transaction.updateItem<GameWorkloadRenderItem>(_spaceRenderItemID, [proxies](GameWorkloadRenderItem& item) {
         item.setAllProxies(proxies);
